Refactor all to typescript and extract hardcoded values to constants
