+++ react/components/Conditions/Separator.tsx
@@ -1,19 +1,24 @@
 import React from 'react'
+import PropTypes from 'prop-types'
 
-interface Props {
- label: string
+const SEPARATOR_LINE_OPTICAL_COMPENSATION = '32px'
+
+const propTypes = {
+  label: PropTypes.string
 }
 
-export default function Separator<Props>({ label }) {
+type Props = PropTypes.InferProps<typeof propTypes>
+
+const Separator: React.FC<Props> = ({ label }) => {
   return (
-    <div>
-      <div
-        style={{ width: 'calc(100% + 34px)' }}
-        className="flex flex-row w-100 nowrap items-center mv3 nl5">
-        <hr className="ma0 b--black-10 bb bb-0 w-50" />
-        <span className="gray ph3 dib bg-white">{label}</span>
-        <hr className="ma0 b--black-10 bb bb-0 w-50" />
-      </div>
+    <div
+      style={{ width: `calc(100% + ${SEPARATOR_LINE_OPTICAL_COMPENSATION})` }}
+      className="flex flex-row w-100 nowrap items-center mv3 nl5">
+      <hr className="ma0 b--black-10 bb bb-0 w-50" />
+      {label && (<span className="gray ph3 dib bg-white">{label}</span>)}
+      <hr className="ma0 b--black-10 bb bb-0 w-50" />
     </div>
   )
 }
+
+export default Separator
